Refactor terminal to fix problem with discarded chars
Termion reads the Read two bytes at a time. If only the first byte is consumed when calling terminal.read_key() and there are more in the buffer, the leftover byte is stored in an Option inside the Keys object.

By the end of read_key, the Keys object is dropped and on the next call, we can not recover this byte.

This PR fixes it by storing the Keys object in the Terminal