fix: store current written output in an in-memory buffer to properly count lines rendered
when the number of lines actually rendered is larger than the ones counted by inquire, for example when user-provided prompt messages have \n characters, the rendering gets all weird because we do not properly clean all lines. We follow a common approach from other libs that is to store the current output and then count the \n characters there present.

Helps fix #15