Refactor architecture of prompts for better key handling (#126)
Major architecture changes:

1. Introduction of Action enums as an intermediary layer between Key events and the prompt's business logic.

There's a top level Action where one of its variants is Inner, with a member of a generic type.

2. Introduction of Config structure for each prompts.

This is still very underutilized. I needed to export some fields to an external struct for proper key to action handling in some cases that depended on the prompt config (e.g. vim key bindings)

In the future, it might be worthwhile to explore using solely this struct for prompt specific fields and then moving shared behavior to a common Prompt struct.

3. Prompt trait

All prompts implement this top level trait which implements a couple of methods that were basically duplicated code before.

4. Input actions

The input struct now also handles actions instead of key events.

---

Future steps after this design change will include exposing calling prompts with a predefined list of actions, instead of relying on reading keys from the terminal.